create UserData context and refactor Auth context
