+++ src/app/layout.tsx
@@ -1,8 +1,9 @@
 import '@/app/globals.css';
 import { Inter } from 'next/font/google';
-import { AuthProviderWrapper } from '@/app/auth.context';
+import { AuthProviderWrapper } from '@/app/context/auth.context';
 import clsx from 'clsx';
 import SideBar from '@/components/SideBar';
+import { UserDataProviderWrapper } from '@/app/context/userData.context';
 
 const inter = Inter({ subsets: ['latin'] });
 
@@ -16,10 +17,12 @@ export default async function RootLayout({ children }: { children: React.ReactNo
     <html lang='en'>
       <body className={clsx(inter.className, 'bg-slate-700 p-4')}>
         <AuthProviderWrapper>
-          <div className='max-w-[320px] mx-auto'>
-            <SideBar />
-            {children}
-          </div>
+          <UserDataProviderWrapper>
+            <div className='max-w-[320px] mx-auto'>
+              <SideBar />
+              {children}
+            </div>
+          </UserDataProviderWrapper>
         </AuthProviderWrapper>
       </body>
     </html>
